requirements.txt
openai
streamlit
faiss-cpu
python-dotenv